--- a/include/linux/blkdev_compat.h
+++ b/include/linux/blkdev_compat.h
@@ -498,7 +498,11 @@ blk_queue_discard_granularity(struct request_queue *q, unsigned int dg)
 #define	VDEV_HOLDER			((void *)0x2401de7)
 
 #ifndef HAVE_GENERIC_IO_ACCT
-#define	generic_start_io_acct(rw, slen, part)		((void)0)
+static inline void
+generic_start_io_acct(int rw, unsigned long sectors, struct hd_struct *part)
+{
+}
+
 static inline void
 generic_end_io_acct(int rw, struct hd_struct *part, unsigned long start_time)
 {