@tuxoko 2017-02-27
> One possible way to correctly handle this would be to take the required range locks synchronously in zvol_request() prior to dispatching the bio to the taskq. That way we're certain all the operations will be processed in the order in which they were received and still done in parallel when possible. We may however need to take some to ensure this can never deadlock due to pending requests in the taskq.

I have thought of that, just wasn't sure if we need it. But I think your right that we should do this. I don't think there's anyway of deadlock since the lock dependency doesn't change at all.

> This would also be a great opportunity to investigating implementing the zvol as a multi-queue block device. In theory this should also allow us to improve the sustained IOPs rate for a zvol.

I had wanted to do that for a long time, but it would mean we need to support two completely different zvol_request path due to legacy kernel support. So I haven't brought myself to do it.