@behlendorf 2017-02-27
> I don't think there's anyway of deadlock since the lock dependency doesn't change at all.

I can't think of a scenario either since we won't be taking any locks in the taskq function.