@sempervictus 2017-02-28
This PR seems to be hitting the codepath which causes the bug i describe in #4265. I ran the same non sync IO tiotests against a fresh ZVOL created before the upgrade, and recreated after the upgrade, and saw something rather interesting. In one of the test cycles _after_ the upgrade, the linear/random write speed ratio flipped - the linear writes went >2X faster than the random, but the random slowed down to hell. I have Linux prefetch disabled on ZVOLs using fad4380317, both before and after, and up till now have not seen anything since the ZVOL overhaul which made linear writes work even near the speed of the underlying SSD.

As i mention in the issue, the ZVOL use case is rather important for shared storage, and today, its impractical for significantly loading consumers and/or contending ones. If there is something serializing the write requests into blocking IOs, this PR might be hitting it, and while the results arent great from the paltry bench i ran (random writes @ half the speed, sometimes, maybe), it seems to be in the right ballpark.

Here's what i'm seeing using the same non-sync-IO tiotest i ran for the issue benchmarks:
Pre patch:
```
| Write        1024 MBs |    6.8 s | 151.517 MB/s |   8.0 %  | 767.9 % |
| Random Write  469 MBs |    1.9 s | 241.722 MB/s |  24.2 %  | 756.3 % |
| Write        1024 MBs |    7.5 s | 137.415 MB/s |   7.7 %  | 711.5 % |
| Random Write  469 MBs |    2.0 s | 234.369 MB/s |  19.4 %  | 751.2 % |
| Write        1024 MBs |    7.5 s | 136.030 MB/s |   7.1 %  | 703.2 % |
| Random Write  469 MBs |    3.0 s | 157.182 MB/s |  12.9 %  | 524.1 % |
```
Post patch:
```
| Write        1024 MBs |    7.5 s | 135.746 MB/s |  11.2 %  | 484.5 % |
| Random Write  469 MBs |    1.8 s | 265.354 MB/s |  23.2 %  | 261.1 % |
| Write        1024 MBs |    3.5 s | 289.724 MB/s |  17.1 %  | 477.1 % |
| Random Write  469 MBs |    3.8 s | 122.846 MB/s |   8.1 %  | 265.2 % |
| Write        1024 MBs |    7.0 s | 145.581 MB/s |   9.5 %  | 249.2 % |
| Random Write  469 MBs |    2.3 s | 199.482 MB/s |  15.1 %  | 176.1 % |
```