@sempervictus 2017-02-28
Running these workloads at 15% of the volsize 10X (each overwriting the last, supposedly), the performance goes down to 80/80MB for writes, while reads appear capped at that ~250 mark. 

On a related note, we need a consistent benchmarking approach which accounts for all the desirable synthetic counters (fio, others?) and real-world applications written by people who dont think about IO models or OS semantics (tiotest, others?), and produces common outputs we can all work with. Different users will be interested in different representations of said benchmarks and their associated tunables, since people backing ASM with ZVOL and those backing Cinder will probably not have the same IO patterns (till a Cinder user builds a RAC in their cloud).

@tuxoko: This work is obviously of great interest to us, so if you need us to test ideas/options, we can definitely allocate some resources, including bare metal if need be.