@tuxoko 2017-02-28
@sempervictus 
Thanks, for the testing. That's strange, I wouldn't expect it would affect sequential read that much.

Also, I'm still struggling with some performance inconsistencies. I originally test this patch on a bare metal 8 spindle disks mirror pair pool with 256GB ram. Doing fio 4k randwrite with 32 iodepth would cause strange behavior as it would write for 3~5 sec and hang for 0.5~2min and repeat. I figured out that reducing zfs_dirty_data_max from 10% total mem to 1G greatly reduce such behavior.

However, when testing on a KVM with ssd back virual disks with 8GB ram. Doing the same test would still have intermittent hang for 2~4sec.