@sempervictus 2017-02-28
@tuxoko: the readahead patch was not helping me here, dropped that, added TRIM support, and numbers are looking a lot better on the same 840 SSD.
Reads after TRIM + allowing Linux readahead ot work on ZVOLs:
```
| Read         1024 MBs |    0.7 s | 1394.343 MB/s |  79.6 %  | 968.1 % |
| Random Read   469 MBs |    2.0 s | 232.904 MB/s |  27.5 %  | 426.8 % |
| Read         1024 MBs |    0.7 s | 1372.518 MB/s |  65.8 %  | 967.8 % |
| Random Read   469 MBs |    2.0 s | 229.402 MB/s |  31.0 %  | 420.7 % |
| Read         1024 MBs |    0.7 s | 1391.270 MB/s |  74.3 %  | 972.9 % |
| Random Read   469 MBs |    2.0 s | 234.224 MB/s |  27.7 %  | 425.2 % |
```
Writes in the same conditions:
```
| Write        1024 MBs |    3.2 s | 319.611 MB/s |  20.4 %  | 459.8 % |
| Random Write  469 MBs |    3.6 s | 128.858 MB/s |  10.9 %  | 131.5 % |
| Write        1024 MBs |    3.6 s | 286.579 MB/s |  15.5 %  | 485.8 % |
| Random Write  469 MBs |    1.2 s | 407.385 MB/s |  23.3 %  | 479.7 % |
| Write        1024 MBs |    3.7 s | 273.846 MB/s |  11.6 %  | 430.9 % |
| Random Write  469 MBs |    0.9 s | 496.128 MB/s |  39.7 %  | 588.1 % |
```
To see how reuse of the volume space affects this stack, i ran the same test 20X with 8 threads instead of 4 to double up on the amount written. Its a bit "all over the place" in terms of numbers, which is slightly concerning, but definitely shows improvement, and doesnt have a clear trend of degrading the performance down to a crawl. Results paste - http://pastebin.com/BvQ9XLnu