@behlendorf 2017-02-28
It may be worth exploring creating multiple zvol taskqs.  Previously there was some concern over contention of the taskq spin lock which could result in inconsistent behavior.  One per multi-queue might work well.

On a related note setting `maxalloc` to the number of threads might be helpful to restrict how fast new tasks can be added to the taskq.  If this approach seems promising we'd want to improve the blocking code in the spl's taskq_alloc() function.