@sempervictus 2017-03-15
So looks like i found some viable tuning knobs in sysfs relating to how the block device writes are handled with this patch:
1 - Linear writes via caching strategy - /sys/devices/virtual/block/zdX/queue/write_cache going from 'write back' to 'write through' using this patch roughly _doubles_ the linear write performance on a used-and-abused ZVOL from ~180MB/s to ~400MB/s. Makes sense, we already have ARC, why write-back?
2 - Random writes by eliminating merging at the Linux layer - /sys/devices/virtual/block/zdX/queue/nomerges going from 0 to 1 took random writes from ~250MB/s to ~700MB/s. Also makes sense since the ZFS pipeline will aggregate/merge those IOs anyway. Another 40-80MB/s in random writes seems to come from setting rotational to 1, which i'm at a loss to explain.
We may want to set ZVOLs to be 'write through' by default with nomerge at 1 if @dweeezil or others can reproduce these results.  The numbers seem to "merit checking out."